requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-telegram-bot>=21.0
openai>=1.3.7
mysql-connector-python>=8.2.0
httpx>=0.25.2
Pillow>=10.2.0
pytz>=2023.3
python-dateutil>=2.8.2